+++ lib/rspec-rerun/rspec.rb
@@ -1,8 +1,0 @@
-module RSpec
-  module Rerun
-    def self.rspec3?
-      require 'rspec'
-      ::RSpec.const_defined?(:Support)
-    end
-  end
-end